+++ Source/ACE.Server/WorldObjects/Player_Xp.cs
@@ -13,8 +13,8 @@ namespace ACE.Server.WorldObjects
     partial class Player
     {
 
-        private static double XpModifier { get { return PropertyManager.GetFloat("xp_modifier"); } }
-        private static double LuminanceModifier { get { return PropertyManager.GetFloat("luminance_modifier"); } }
+        private static double XpModifier { get { return PropertyManager.GetDouble("xp_modifier"); } }
+        private static double LuminanceModifier { get { return PropertyManager.GetDouble("luminance_modifier"); } }
 
         /// <summary>
         /// Raise the available XP by a specified amount
